requirements.txt
streamlit>=1.30.0
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyarrow>=12.0.0
plotnine>=0.12.0
